Debug test code added to INKEY
I would appreciate a test build with this. It adds as test code INKEY(&8000+nn) to do a raw SDL keytest, so then I can work out real-world key responses to get the negative INKEY lookup table correct. This is the equivalent of INKEY(&FE00+nn) in the console module, but SDL keycodes are larger than 8 bits. (console uses INKEY(&8000+nn) to return a 16-bit keypress.)