Updates to INKEY functions
I've been dragging a stiff comb through keyboard.c integrating all the
scattered code.

If NEWKBD is not defined it builds with the old keyboard handling code
and is functionally unchanged (and should be binary identical).

If NEWKBD is defined it builds with the new keyboard handling code.

All the platform-specific functions duplicate code which makes them more
appropriate to be single functions with bits of platform-specific code.
It also is gradually eliminating the BODGE identifiers I've had to put in
to get some of the builds working.

The majority of the INKEY code is now integrated and just depends on a
few bits of legacy code. Some of the legacy INKEY routines didn't even
support INKEY-256, so the first result was that all platforms support
INKEY-256 automatically.

+ve INKEY tested and works on: RISCOS:ok, WinSDL:ok, MGW:ok, DJP:ok
-ve INKEY tested and works on: RISCOS:ok, WinSDL:ok, MGW:ok, DJP:to do

The SDL builds now support testing for keys not known about by SDL by
using helper code for a handful of keyscan values.

DOS/Win builds without SDL now support negative INKEYs using code
from the JGH 'console' library.