Add `--` command-line option
This is a common construct in other utilities, and indicates that any
subsequent arguments should not be consumed as options to the interpreter.
Since Brandy swallows all arguments that start (case-insensitively) with
the first one or two characters of options that it recognises, it is otherwise
impossible to correctly execute BASIC programs which themselves support
options which also match one of these patterns.

For example, a program can't have a `-quiet` option because it is confused
with `-quit`:

  brandy -chain MyProg -quiet
  No filename was supplied after option '-quiet'

but with this new facility, we can write:

  brandy -chain MyProg -- -quiet