Prevent fall-through in switch/case
This fixes compiler warnings.

We appear to fall-through when we get an invalid two byte token. I'm not
convinced the current behaviour is correct. I suspect we should be reporting
an error and rejecting the file, however I don't have a large pile of
tokenised BBC BASIC from the wild to test such a change against.